+++ Packages/com.quickeye.ui-toolkit-plus/Editor/QuickEye.UxmlBridgeGen/UxmlParser.cs
@@ -4,6 +4,8 @@ using System.Xml.Linq;
 
 namespace QuickEye.UxmlBridgeGen
 {
+    using UnityEngine;
+
     internal static class UxmlParser
     {
         /// <summary>